requirements.txt
ultralytics>=8.3.0,<9
lxml>=4.9
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
opencv-python-headless>=4.8
pycocotools>=2.0.7
pillow>=10.0
pyyaml>=6.0
